Added examples and removed tests directory
